Refactor keep consts for values, move helpers to libs
